requirement.txt
streamlit
google-cloud-speech
pyaudio